+++ examples/counter_app/src/main.rs
@@ -2,7 +2,7 @@ use crossterm::event::KeyCode;
 use ratatui::Frame;
 use ratatui::widgets::{Block, Paragraph};
 use ratatui_router::ratatui_router_derive::Routes;
-use ratatui_router::router::{Events, Routed, Router};
+use ratatui_router::router::{EventHook, Events, Routed, Router};
 #[derive(Routes)]
 pub enum MyRoutes {
     Home { counter: i64 },
@@ -11,16 +11,26 @@ pub enum MyRoutes {
 
 pub fn Home(ctx: &mut Router<MyRoutes>, frame: &mut Frame, counter: &mut i64) {
     let modifier = ctx.get_context::<i64>();
-    if let Some(ev) = ctx.event() {
-        match ev {
-            Events::Event(crossterm::event::Event::Key(key)) => match key.code {
-                KeyCode::Esc => ctx.exit(),
-                KeyCode::Tab => ctx.change_page(MyRoutes::Modifier),
-                _ => *counter += 1 * *modifier.borrow(),
-            },
-            _ => {}
-        }
-    }
+    // if let Some(ev) = ctx.get_event() {
+    //     match ev {
+    //         Events::Event(crossterm::event::Event::Key(key)) => match key.code {
+    //             KeyCode::Esc => ctx.exit(),
+    //             KeyCode::Tab => ctx.change_page(MyRoutes::Modifier),
+    //             _ => *counter += 1 * *modifier.borrow(),
+    //         },
+    //         _ => {}
+    //     }
+    // }
+    // as alternative
+    ctx.use_event(|ctx, ev| match ev {
+        crossterm::event::Event::Key(key) => match key.code {
+            KeyCode::Esc => ctx.exit(),
+            KeyCode::Tab => ctx.change_page(MyRoutes::Modifier),
+            _ => *counter += 1 * *modifier.borrow(),
+        },
+        _ => {}
+    });
+    
     let paragraph = Paragraph::new(
         format!("Current counter = {}, press escape to exit, tab to change page, or any other button to increment to {}", counter, *modifier.borrow())
     )
@@ -30,17 +40,15 @@ pub fn Home(ctx: &mut Router<MyRoutes>, frame: &mut Frame, counter: &mut i64) {
 
 pub fn Modifier(ctx: &mut Router<MyRoutes>, frame: &mut Frame) {
     let modifier = ctx.get_context::<i64>();
-    if let Some(ev) = ctx.event() {
-        match ev {
-            Events::Event(crossterm::event::Event::Key(key)) => match key.code {
-                KeyCode::Esc => ctx.exit(),
-                KeyCode::Tab => {
-                    ctx.go_back();
-                }
-                _ => *modifier.borrow_mut() += 1,
-            },
-            _ => {}
-        }
+    match ctx.get_event() {
+        Events::Event(crossterm::event::Event::Key(key)) => match key.code {
+            KeyCode::Esc => ctx.exit(),
+            KeyCode::Tab => {
+                ctx.go_back();
+            }
+            _ => *modifier.borrow_mut() += 1,
+        },
+        _ => {}
     }
     let paragraph = Paragraph::new(
         format!("Current modifier = {}, press escape to exit, tab to change page, or any other button to increment", *modifier.borrow())